requirements.txt
numpy==1.21.0
opencv_python==4.5.3.56
pydicom==2.1.2
tensorflow-gpu==2.3.0
protobuf==3.17.3
PyYAML==5.4.1
split_folders==0.4.3
